Docker_Game_Kadai_DockerProject/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
Pillow==10.1.0
